:recycle: api: check language availability in certif course instead of user
