+++ api/lib/domain/usecases/index.js
@@ -52,6 +52,7 @@ import * as missionRepository from '../../../src/school/infrastructure/repositor
 import * as schoolRepository from '../../../src/school/infrastructure/repositories/school-repository.js';
 import * as codeGenerator from '../../../src/shared/domain/services/code-generator.js';
 import * as cryptoService from '../../../src/shared/domain/services/crypto-service.js';
+import * as languageService from '../../../src/shared/domain/services/language-service.js';
 import { tokenService } from '../../../src/shared/domain/services/token-service.js';
 import * as adminMemberRepository from '../../../src/shared/infrastructure/repositories/admin-member-repository.js';
 import * as answerRepository from '../../../src/shared/infrastructure/repositories/answer-repository.js';
@@ -294,6 +295,7 @@ const dependencies = {
   learningContentConversionService,
   learningContentRepository,
   localeService,
+  languageService,
   mailService,
   membershipRepository,
   missionRepository,
